+++ packages/leavitt-elements/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 1.7.4 (2021-08-27)
+
+**Note:** Version bump only for package @leavittsoftware/leavitt-elements
+
+
+
+
+
+## 1.7.3 (2021-08-19)
+
+**Note:** Version bump only for package @leavittsoftware/leavitt-elements
+
+
+
+
+
 ## 1.7.2 (2021-07-29)
 
 **Note:** Version bump only for package @leavittsoftware/leavitt-elements